[CHORE] #60 - Database 수정
