+++ app/src/main/java/com/ivyclub/contact/model/PhoneContactData.kt
@@ -1,0 +1,6 @@
+package com.ivyclub.contact.model
+
+data class PhoneContactData(
+    val name: String,
+    val phoneNumber: String
+)
