--- app/src/main/java/com/ivyclub/contact/ui/main/friend/FriendListViewType.kt
@@ -1,5 +1,0 @@
-package com.ivyclub.contact.ui.main.friend
-
-enum class FriendListViewType {
-    GROUP_NAME, FRIEND, GROUP_DIVIDER
-}
